+++ ReleaseNotes.md
@@ -9,7 +9,7 @@ Please consult the [database upgrade documentation](https://eoscta.docs.cern.ch/
 
 ### Features
 - cta/CTA#1163 - cta-admin now prefixes the drivename with a '!' if the respective logical library is disabled
-
+- cta/CTA#1168 - Add configuration option for scheduler stack size to /etc/cta/cta-frontend-xrootd.conf
 ### Bug fixes
 - cta/CTA#1156 - failed to instantiate the RAO algorithm